kpi_tracker_bot/requirements.txt
python-telegram-bot==20.7
aiosqlite
pandas
numpy
matplotlib
seaborn
APScheduler
python-dotenv
openpyxl
